Added summary comments to the DataLocation enum for future use. Moved DataLoader files out of the DataLoader folder/namespace for easier use.
