--- FileUtilities/DataLoader/DataLocation.cs
@@ -1,11 +1,0 @@
-namespace FluffyVoid.FileUtilities.DataLoader;
-
-/// <summary>
-///     The types of data storage locations the data loader supports
-/// </summary>
-public enum DataLocation
-{
-    ApplicationData,
-    UserData,
-    SaveData
-}
